% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sl}
\alias{sl}
\title{Smoothed LASSO Laplacian}
\usage{
sl(p)
}
\arguments{
\item{p}{Dimension of the matrix.}
}
\value{
A square matrix with dimension \code{p} by \code{p}. The matrix used in the quadratic form of the smoothed type of penalty.
}
\description{
This is the function that used for creating 1 dimensional smoothed (squared difference) matrix.
}
\examples{
sl(5)
}
